+++ cs50/Week 2/argv-2.c
@@ -12,7 +12,7 @@
 
 int main(int argc, string argv[]) {
 	for (int i = 0; i < argc; i++) {
-		for (int j = 0, n = strlen(argv[i]); j < n; j++) {
+		for (int j = 0, n = (int)strlen(argv[i]); j < n; j++) {
 			printf("%c\n", argv[i][j]);
 		}
 	}
