+++ cs50/Week 2/capitalize-0.c
@@ -12,7 +12,7 @@
 
 int main(void) {
 	string s = GetString();
-	for (int i = 0, n = strlen(s); i < n; i++) {
+	for (int i = 0, n = (int)strlen(s); i < n; i++) {
 		if (s[i] >= 'a' && s[i] <= 'z') {
 			printf("%c", s[i] - ('a' - 'A'));
 		} else {
