+++ cs50/Week 2/capitalize-1.c
@@ -13,7 +13,7 @@
 
 int main(void) {
 	string s = GetString();
-	for (int i = 0, n = strlen(s); i < n; i++) {
+	for (int i = 0, n = (int)strlen(s); i < n; i++) {
 		if (islower(s[i])) {
 			printf("%c", toupper(s[i]));
 		} else {
