+++ cs50/Week 2/string-0.c
@@ -13,7 +13,7 @@
 int main(void) {
 	string s = GetString();
 	
-	for (int i = 0; i < strlen(s); i++) {
+	for (int i = 0; i < (int)strlen(s); i++) {
 		printf("%c\n", s[i]);
 	}
 }
