+++ cs50/Week 2/string-1.c
@@ -14,7 +14,7 @@ int main(void) {
 	string s = GetString();
 	
 	if (s != NULL) {
-		for (int i = 0; i < strlen(s); i++) {
+		for (int i = 0; i < (int)strlen(s); i++) {
 			printf("%c\n", s[i]);
 		}
 	}
