+++ cs50/Week 2/string-2.c
@@ -14,7 +14,7 @@ int main(void) {
 	string s = GetString();
 	
 	if (s != NULL) {
-		for (int i = 0, n = strlen(s); i < n; i++) {
+		for (int i = 0, n = (int)strlen(s); i < n; i++) {
 			printf("%c\n", s[i]);
 		}
 	}
